requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.13
psycopg2-binary>=2.9
djangorestframework-simplejwt